requirements.txt
fastapi==0.111.0
SQLAlchemy==2.0.29
alembic==1.13.1
python-slugify==8.0.4
asyncpg==0.29.0